requirements.txt
asgiref==3.5.0
astroid==2.4.2
autopep8==1.5.4
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.3
isort==5.6.4
itypes==1.2.0
Jinja2==3.0.3
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.2.0
Pillow==9.0.1
pycodestyle==2.6.0
pycparser==2.21
PyJWT==2.3.0
pylint==2.6.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
wincertstore==0.2
wrapt==1.12.1
